Deduplicate snapshot merge utility files
Share GitTool and process/file utility implementations between SnapshotTesting and InlineSnapshotTesting via common linked sources, and add a SnapshotTesting regression test for GitTool command parsing.
