+++ tests/Meziantou.Framework.SnapshotTesting.Tests/SnapshotTests.cs
@@ -1,6 +1,7 @@
 #nullable enable
 
 using System.Text.RegularExpressions;
+using Meziantou.Framework.SnapshotTesting.MergeTools;
 
 namespace Meziantou.Framework.SnapshotTesting.Tests;
 
@@ -400,6 +401,18 @@ public sealed class SnapshotTests
         };
 
         ValidateWithSerializerCount(validateSettings, count: 2);
+    }
+
+    [Theory]
+    [InlineData("ping", "ping", "")]
+    [InlineData("ping ", "ping", "")]
+    [InlineData("ping a b", "ping", "a b")]
+    [InlineData("\"ping\"", "ping", "")]
+    [InlineData("\"ping\" ", "ping", "")]
+    [InlineData("\"ping\" a b", "ping", "a b")]
+    public void GitTool_ParseCommand(string value, string command, string arguments)
+    {
+        Assert.Equal((command, arguments), GitTool.ParseCommandFromConfiguration(value));
     }
 
     private static void ValidateWithSerializerCount(SnapshotSettings settings, int count)
